+++ src/components/_common/image/XpImage.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { MediaType, XpImageProps } from '../../../types/media';
+import { getMediaUrl } from '../../../utils/urls';
 
 type Props = {
     imageProps: XpImageProps;
@@ -8,16 +9,17 @@ type Props = {
     className?: string;
 };
 
-export const getImageUrl = (imageProps: XpImageProps, scale?: string) => {
-    if (!imageProps) {
+export const getImageUrl = (image: XpImageProps, scale?: string) => {
+    if (!image) {
         return null;
     }
 
-    if (imageProps.__typename === MediaType.Image && scale) {
-        return imageProps.imageUrl?.replace('$scale', scale);
-    }
+    const url =
+        image.__typename === MediaType.Image && scale
+            ? image.imageUrl?.replace('$scale', scale)
+            : image.mediaUrl;
 
-    return imageProps.mediaUrl;
+    return getMediaUrl(url);
 };
 
 export const XpImage = ({ imageProps, alt = '', scale, className }: Props) => {
