+++ src/utils/urls.ts
@@ -30,6 +30,9 @@ const nofollowPattern = new RegExp(`^(${appOrigin})?(\\/sok($|\\?|\\/))`, 'i');
 
 export const isNofollowUrl = (url: string) => nofollowPattern.test(url);
 
+export const isNavUrl = (url: string) =>
+    /^(https:\/\/)?([a-z0-9-]*(\.)?nav\.no)?(\/|$)/i.test(url);
+
 export const stripXpPathPrefix = (path: string) =>
     path?.startsWith(xpContentPathPrefix)
         ? path.slice(xpContentPathPrefix.length)
